+++ action/action-util/src/main/java/com/powsybl/action/util/Scalable.java
@@ -19,11 +19,11 @@ import java.util.stream.Collectors;
  */
 public interface Scalable {
 
-    float initialValue(Network n);
+    double initialValue(Network n);
 
     void reset(Network n);
 
-    float maximumValue(Network n);
+    double maximumValue(Network n);
 
     void listGenerators(Network n, List<Generator> generators, List<String> notFoundGenerators);
 
@@ -31,7 +31,7 @@ public interface Scalable {
 
     List<Generator> listGenerators(Network n);
 
-    float scale(Network n, float asked);
+    double scale(Network n, double asked);
 
     static GeneratorScalable gen(String id) {
         return new GeneratorScalable(id);
