+++ iidm/iidm-api/src/main/java/com/powsybl/iidm/network/TwoTerminalsConnectable.java
@@ -23,7 +23,7 @@ public interface TwoTerminalsConnectable<I extends TwoTerminalsConnectable<I>> e
 
         CurrentLimits.TemporaryLimit getTemporaryLimit();
 
-        float getPreviousLimit();
+        double getPreviousLimit();
 
         String getPreviousLimitName();
     }
